refactor(Clipboard copy related components): review comments
